Refactoring of the linked list implementation.
Hey Priya!

A few notes.

It is generally bad practice to define functions that will be used by all instances of an object on the constructor (you want to define it on the prototype instead, so there is a shared reference to all those functions VS one reference being held for each generic function by each instance you create).

Instead of iterating through the array each time you want to get the length of the linked list, maybe just keep a property on the list object? 

Finally, tests are always a plus, and not too hard to write, so if you can, always write tests.

Let me know if I can clarify any of these changes! Happy coding!